adds more cleanup
