+++ insecure/wintermute/attackOrchestrator_test.go
@@ -334,31 +334,6 @@ func orchestratorOutputSanityCheck(
 	require.Equal(t, expBouncedReceiptCount, actualBouncedReceiptCount)
 }
 
-func chunkDataPackRequestsSanityCheck(t *testing.T,
-	outputEvents []*insecure.Event, // list of all output events of the wintermute orchestrator.
-	originalChunkRequestIds flow.IdentifierList, // list of original chunk data requests sent by corrupted verification nodes.
-	expectedBouncedBackReqs int, // number of expected bounced back chunk data pack requests.
-) {
-
-	bouncedBackRequestedChunkIds := flow.IdentifierList{}
-
-	for _, outputEvent := range outputEvents {
-		switch event := outputEvent.FlowProtocolEvent.(type) {
-		case *messages.ChunkDataRequest:
-			bouncedBackRequestedChunkIds = bouncedBackRequestedChunkIds.Union(flow.IdentifierList{event.ChunkID})
-		}
-	}
-
-	// number of bounced receipts should match the expected value.
-	actualBouncedChunkDataRequestCount := 0
-	for _, chunkId := range bouncedBackRequestedChunkIds {
-		if originalChunkRequestIds.Contains(chunkId) {
-			actualBouncedChunkDataRequestCount++
-		}
-	}
-	require.Equal(t, expectedBouncedBackReqs, actualBouncedChunkDataRequestCount)
-}
-
 // receiptsWithDistinctResultFixture creates a set of execution receipts (with distinct result) one per given executor id.
 // It returns a map of execution receipts to their relevant attack network events.
 func receiptsWithDistinctResultFixture(
@@ -389,97 +364,6 @@ func receiptsWithDistinctResultFixture(
 
 	require.Len(t, eventMap, count*len(exeIds))
 	return eventMap, receipts
-}
-
-// receiptsWithSameResultFixture creates a set of receipts (all with the same result) per given executor id.
-// It returns a map of execution receipts to their relevant attack network events.
-func receiptsWithSameResultFixture(
-	t *testing.T,
-	count int, // total receipts per execution id.
-	exeIds flow.IdentifierList, // identifier of execution nodes.
-	targetIds flow.IdentifierList, // target recipients of the execution receipts.
-) (map[flow.Identifier]*insecure.Event, []*flow.ExecutionReceipt) {
-	// list of execution receipts
-	receipts := make([]*flow.ExecutionReceipt, 0)
-
-	// map of execution receipt ids to their event.
-	eventMap := make(map[flow.Identifier]*insecure.Event)
-
-	// generates "count"-many receipts per execution nodes with the same
-	// set of results.
-	for i := 0; i < count; i++ {
-
-		result := unittest.ExecutionResultFixture()
-
-		for _, exeId := range exeIds {
-			receipt := unittest.ExecutionReceiptFixture(
-				unittest.WithExecutorID(exeId),
-				unittest.WithResult(result))
-
-			require.Equal(t, result.ID(), receipt.ExecutionResult.ID())
-
-			event := executionReceiptEvent(receipt, targetIds)
-
-			_, ok := eventMap[receipt.ID()]
-			require.False(t, ok) // check for duplicate receipts.
-
-			receipts = append(receipts, receipt)
-			eventMap[receipt.ID()] = event
-		}
-	}
-
-	require.Len(t, eventMap, count*len(exeIds))
-	return eventMap, receipts
-}
-
-// executionReceiptEvent creates the attack network event of the corresponding execution receipt.
-func executionReceiptEvent(receipt *flow.ExecutionReceipt, targetIds flow.IdentifierList) *insecure.Event {
-	return &insecure.Event{
-		CorruptedId:       receipt.ExecutorID,
-		Channel:           engine.PushReceipts,
-		Protocol:          insecure.Protocol_UNICAST,
-		TargetIds:         targetIds,
-		FlowProtocolEvent: receipt,
-	}
-}
-
-// chunkDataPackResponseForReceipts creates and returns chunk data pack response as well as their corresponding events for the given set of receipts.
-func chunkDataPackResponseForReceipts(receipts []*flow.ExecutionReceipt, verIds flow.IdentifierList) ([]*insecure.Event, flow.IdentifierList) {
-	chunkIds := flow.IdentifierList{}
-	responseList := make([]*insecure.Event, 0)
-
-	for _, receipt := range receipts {
-		result := receipt.ExecutionResult
-		for _, chunk := range result.Chunks {
-			chunkId := chunk.ID()
-
-			if chunkIds.Contains(chunkId) {
-				// chunk data pack request already created
-				continue
-			}
-
-			cdpRep := &messages.ChunkDataResponse{
-				ChunkDataPack: *unittest.ChunkDataPackFixture(chunkId),
-			}
-			chunkIds = chunkIds.Union(flow.IdentifierList{chunkId})
-
-			// creates a request event per verification node
-			for _, verId := range verIds {
-				event := &insecure.Event{
-					CorruptedId:       receipt.ExecutorID,
-					Channel:           engine.RequestChunks,
-					Protocol:          insecure.Protocol_PUBLISH,
-					TargetNum:         0,
-					TargetIds:         flow.IdentifierList{verId},
-					FlowProtocolEvent: cdpRep,
-				}
-
-				responseList = append(responseList, event)
-			}
-		}
-	}
-
-	return responseList, chunkIds
 }
 
 // TestRespondingWithCorruptedAttestation evaluates when the Wintermute orchestrator receives a chunk data pack request from a CORRUPTED
